cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "marcyra"
requires-python = ">=3.13"
version = "0.1.0"
dependencies = [
    "pillow",
    "materialyoucolor",
    "numpy"
]

[project.scripts]
marcyra = "marcyra:main"

[tool.hatch.build.targets.sdist]
only-include = [
    "src"
]

[tool.ruff]
line-length = 120
src = ["src"]

[tool.ruff.lint.isort]
known-third-party = ["materialyoucolor", "PIL", "numpy"]

